+++ components.js
@@ -2,7 +2,7 @@
 
 import { TWChatInput, TWChatButtonsContainer, TWChatSendButton } from "./tw_elements_finder.js"
 import { popupButtonForEditor } from "./code_editor.js"
-import { previewMath } from "./render_math.js"
+import { previewMath } from "./math/render_math.js"
 import { updateBuffer, sendMessage, prevMessage, nextMessage } from "./history.js"
 
 function createPreviewBox() {
